extracted/cleaned up community and post types
