--- client/components/ViewPost/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./viewpost";
