+++ client/pages/r/[communityName]/[postId]/index.tsx
